+++ backend/core-api/src/modules/broadcast/db/definitions/engages.ts
@@ -20,7 +20,8 @@ export const engageMessageSchema = new Schema(
 
     title: { type: String, label: 'Title', required: true },
     cpId: { type: String, label: 'Client portal ID' },
-    fromUserId: { type: String, label: 'From user', required: true },
+    fromEmail: { type: String, label: 'From sender' },
+    fromUserId: { type: String, label: 'From user' },
 
     method: {
       type: String,