+++ frontend/libs/ui-modules/src/modules/products/hooks/useProducts.ts
@@ -3,7 +3,7 @@ import {
   GET_ASSIGNED_PRODUCTS,
   GET_PRODUCTS,
 } from '../graphql/queries/productsQueries';
-import { QueryHookOptions, useQuery } from '@apollo/client';
+import { NetworkStatus, QueryHookOptions, useQuery } from '@apollo/client';
 
 import { IProduct } from '../types/Product';
 
@@ -11,19 +11,24 @@ const PRODUCTS_LIMIT = 30;
 export const useProducts = (
   options?: QueryHookOptions<ICursorListResponse<IProduct>>,
 ) => {
-  const { data, loading, fetchMore, error } = useQuery<
+  const { data, loading, fetchMore, error, networkStatus } = useQuery<
     ICursorListResponse<IProduct>
   >(GET_PRODUCTS, {
     ...options,
+    notifyOnNetworkStatusChange: true,
     variables: {
       limit: PRODUCTS_LIMIT,
       ...options?.variables,
     },
   });
   const { list = [], totalCount = 0, pageInfo } = data?.productsMain || {};
+  const fetchingMore = networkStatus === NetworkStatus.fetchMore;
 
   const handleFetchMore = () => {
-    if (!pageInfo || totalCount <= list.length) return;
+    if (!pageInfo || totalCount <= list.length || fetchingMore) {
+      return;
+    }
+
     fetchMore({
       variables: {
         ...options?.variables,
@@ -47,7 +52,8 @@ export const useProducts = (
   };
   return {
     products: list,
-    loading,
+    loading: loading && !fetchingMore,
+    fetchingMore,
     handleFetchMore,
     totalCount,
     error,