+++ frontend/plugins/frontline_ui/src/widgets/automations/modules/facebook/components/action/components/replyMessage/MessageActionForm.tsx
@@ -6,6 +6,7 @@ import {
   useAutomationRemoteFormSubmit,
   useFormValidationErrorHandler,
 } from 'ui-modules';
+import { getMaxMessagesForTrigger } from '~/widgets/automations/modules/facebook/components/action/constants/ReplyMessage';
 import { FacebookMessages } from '~/widgets/automations/modules/facebook/components/action/components/replyMessage/FacebookMessages';
 import { MessageSequenceHeader } from '~/widgets/automations/modules/facebook/components/action/components/replyMessage/MessageSequenceHeader';
 import { ReplyMessageProvider } from '~/widgets/automations/modules/facebook/components/action/context/ReplyMessageProvider';
@@ -19,6 +20,7 @@ export const MessageActionForm = ({
   formRef,
   currentAction,
   onSaveActionConfig,
+  trigger,
 }: AutomationActionFormProps<TMessageActionForm>) => {
   const form = useForm<TMessageActionForm>({
     resolver: zodResolver(replyMessageFormSchema),
@@ -35,7 +37,10 @@ export const MessageActionForm = ({
   });
 
   return (
-    <ReplyMessageProvider form={form}>
+    <ReplyMessageProvider
+      form={form}
+      maxMessages={getMaxMessagesForTrigger(trigger?.type)}
+    >
       <div className="w-[600px]">
         <MessageSequenceHeader />
         <FacebookMessages />