+++ frontend/plugins/sales_ui/src/pages/SalesSettingsIndexPage.tsx
@@ -24,20 +24,18 @@ const PosEditPage = lazy(() =>
 const DealsSettings = () => {
   const { t } = useTranslation('sales');
   return (
-    <PageContainer className="flex-row">
-      <BoardsList />
-      <div className="flex overflow-hidden relative flex-col flex-1">
-        <SettingsHeader breadcrumbs={[]}>
-          <Button variant="ghost" className="font-semibold">
-            <IconSandbox className="w-4 h-4 text-accent-foreground" />
-            {t('boards-and-pipelines')}
-          </Button>
-          <PipelineFormBar />
-        </SettingsHeader>
-        <div className="flex overflow-hidden flex-auto w-full">
-          <div className="flex overflow-hidden flex-col w-full">
-            <PipelineRecordTable />
-          </div>
+    <PageContainer>
+      <SettingsHeader breadcrumbs={[]}>
+        <Button variant="ghost" className="font-semibold">
+          <IconSandbox className="w-4 h-4 text-accent-foreground" />
+          {t('boards-and-pipelines')}
+        </Button>
+        <PipelineFormBar />
+      </SettingsHeader>
+      <div className="flex flex-auto overflow-hidden">
+        <BoardsList />
+        <div className="relative flex min-w-0 flex-1 flex-col overflow-hidden">
+          <PipelineRecordTable />
         </div>
       </div>
     </PageContainer>